osmet: move copy_n and copy_exactly_n out of osmet
